Separate scene validation logic into a separate module
